requirements.txt
asgiref==3.8.1
attrs==25.3.0
autopep8==2.3.2
black==24.10.0
cfgv==3.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.9
Django==5.1.3
django-cors-headers==4.4.0
django-silk==5.3.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
filelock==3.16.1
flake8==7.1.1
gprof2dot==2024.6.6
gunicorn==23.0.0
identify==2.6.3
inflection==0.5.1
isort==6.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
pre_commit==4.0.1
psycopg2==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
ruff==0.9.4
setuptools==75.8.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
virtualenv==20.28.0
